requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
django-tables2==2.3.1
djangorestframework==3.12.1
filelock==3.0.12
Flask==1.1.2
Flask-SocketIO==4.3.1
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
mutagen==1.45.1
mysqlclient==1.4.6
psycopg2==2.8.5
pycparser==2.20
pycryptodomex==3.9.8
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
spotipy==2.16.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.25
Werkzeug==1.0.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.1.0
